+++ dx_router/CHANGELOG.md
@@ -5,3 +5,7 @@
 ## 1.0.0+1
 
 - Changed README.md
+
+## 1.0.0+2
+
+- Updated example based on [dx_generator]('https://pub.dev/packages/dx_generator') (v2.0.0)